Add context trail to links of all items in trail
Also convert several classes to TS
