+++ README.md
@@ -5,13 +5,14 @@ This repository contains client and server code for the [Howdju](https://www.how
 ## Automation status
 
 - [![CI](https://github.com/Howdju/howdju/actions/workflows/ci.yml/badge.svg?branch=master&event=push)](https://github.com/Howdju/howdju/actions/workflows/ci.yml)
-- [![Deploy to preprod](https://github.com/Howdju/howdju/actions/workflows/deploy-preprod.yml/badge.svg?branch=master&event=workflow_run)](https://github.com/Howdju/howdju/actions/workflows/deploy-preprod.yml)
+- [![Deploy to
+preprod](https://github.com/Howdju/howdju/actions/workflows/deploy-preprod.yml/badge.svg?branch=master&event=workflow_run)](https://github.com/Howdju/howdju/actions/workflows/deploy-preprod.yml)
 
 ## Introduction
 
-[Howdju](https://www.howdju.com) is a platform for analyzing and sharing critical analysis of
-claims using evidence. The content is currently user-generated, and we hope to augment users'
-actions with machine learning.
+[Howdju](https://www.howdju.com) is a platform for analyzing and sharing critical analysis of claims
+using evidence. The content is currently user-generated, and we hope to augment users' actions with
+machine learning.
 
 ## Development
 
@@ -26,24 +27,27 @@ setup for developing Howdju.
 
 You can [join our
 Slack](https://join.slack.com/t/howdju/shared_invite/zt-1qbfzlfsj-YRswgQ5RCLDHelef6ya6xg) to discuss
-Howdju with other useres and the site maintainers.
+Howdju with other users and the site maintainers.
 
 ## Code history
 
 The Howdju code base has undergone a few transformations, and since it is currently a prototype for
 exploring features, we have not thoroughly completed all refactorings/cleanups/migrations. This
-section describes major in-progress projects to provide the context for inconsistencies you may find.
+section describes major in-progress projects to provide the context for inconsistencies you may
+find.
 
 ### TypeScript migration
 
 The code base is a little over half TypeScript now and includes TypeScript examples for most major
-components. See #1 for details on the initial migration effort and [issues in the `ts-migration`
-label](https://github.com/Howdju/howdju/labels/ts-migration) for substantial planned effort.
+components. See Howdju/howdju#1 for details on the initial migration effort and [issues in the
+`ts-migration` label](https://github.com/Howdju/howdju/labels/ts-migration) for substantial planned
+effort.
 
 ### Zod validation
 
 All of Howdju's entities are defined using [Zod](https://zod.dev/). Zod is also our preferred
-validation library. See #26 for the status of migrating previous data validation definitions to Zod.
+validation library. See Howdju/howdju#26 for the status of migrating previous data validation
+definitions to Zod.
 
 ### Justification types and evidence representation
 
@@ -64,15 +68,16 @@ evolution. This list represents the status of some relevant entities:
   because the system need not contain an association between a `Source` and entered evidence.)
 - `WritQuote`: a type of `SourceExcerpt`, along with `PicRegion` and `VidSegment`. In many places we
   assume that a `WritQuote` is the only type of `SourceExcerpt`, but we intend to generalize this.
-  E.g., we will replace `WritQuote`-based `Justification`s with `SourceExcerpt`-based justifications.
+  E.g., we will replace `WritQuote`-based `Justification`s with `SourceExcerpt`-based
+  justifications.
 
 ### UI Framework
 
 The web app currently uses react-md@1. Besides being outdated, we would like to explore a UI
-framework that supports both react-native mobile and web. This cross-platform support would enable reusing
-components on both webpages and mobile app screens.
+framework that supports both react-native mobile and web. This cross-platform support would enable
+reusing components on both webpages and mobile app screens.
 
-See [#304](https://github.com/Howdju/howdju/issues/304).
+See Howdju/howdju#304.
 
 ### Premiser name
 
@@ -83,14 +88,16 @@ is now the preferred name.
 
 If you'd like to contribute to Howdju, great! Please feel free to join our Slack to introduce
 yourself and chat about what interests you. If you're specifically interested in contributing code
-changes to Howdju's source code or issues to this repository, please see our [Contributor guidelines](https://github.com/Howdju/howdju/blob/master/CONTRIBUTING.md).
+changes to Howdju's source code or issues to this repository, please see our [Contributor
+guidelines](https://github.com/Howdju/howdju/blob/master/CONTRIBUTING.md).
 
 ## Code layout
 
 This repository is a monorepo based on Yarn berry workspaces. This diagram shows the high level
 dependencies between the packages:
 
-![package dependency diagram](https://raw.githubusercontent.com/Howdju/howdju/master/docs/diagrams/Howdju%20Monorepo%20Package%20Dependencies.drawio.png?token=GHSAT0AAAAAABYMGSPWSANRYRI5BMRJO35YZAAGEQA)
+![package dependency
+diagram](https://raw.githubusercontent.com/Howdju/howdju/master/docs/diagrams/Howdju%20Monorepo%20Package%20Dependencies.drawio.png?token=GHSAT0AAAAAABYMGSPWSANRYRI5BMRJO35YZAAGEQA)
 
 - `howdju-common`: code common to client and server runtimes. Any package can depend on this
   package. It includes:
@@ -128,20 +135,32 @@ Missing from the diagram are:
 
 The howdju.com web app consists of these parts:
 
+### Web app and lambda API
+
 - An HTML bootstrap page served from an S3 bucket which loads
 - A single JS bundle served from Cloudfront that calls
 - A monolithic AWS lambda behind AWS API Gateway.
 - Persistance is in AWS RDS Postgres.
 
+### Message handler
+
+- Howdju currently has a single SNS-based message handler. It is currently responsible for sending
+  email notifications.
+
+### Scheduled jobs
+
+- Howdju has a few scheduled jobs that calculate aggregate scores on justifications.
+
 ## Code architecture
 
-The following sections briefly discuss the major components and dependencies of our clients and services.
+The following sections briefly discuss the major components and dependencies of our clients and
+services.
 
 ### Web app
 
 The web app uses react-navigation to select pages. The pages dispatch Redux actions handled by Redux
-reducers and redux-saga (for asynchronous handling.) It calls the API using Axios. It normalizes entities using
-normalizr.
+reducers and redux-saga (for asynchronous handling.) It calls the API using Axios. It normalizes
+entities using normalizr.
 
 ### API
 
@@ -158,6 +177,6 @@ development docs.
 ## Affiliations
 
 Carl has been attending meetings semiregularly with the [Canonical Debate
-Lab](https://canonicaldebatelab.com/), a lose assocation of people who are all interested in solving
-similar problems (empowering collective intelligence) but with different focuses and approaches. If
-Howdju sounds interesting to you, that group may also interest you.
+Lab](https://canonicaldebatelab.com/), a loose assocation of people who are all interested in
+solving similar problems (empowering collective intelligence) but with different focuses and
+approaches. If Howdju sounds interesting to you, that group may also interest you.
